Scroll follows selection only when selection goes offscreen

+++ src/novelbrowser.h
@@ -18,7 +18,8 @@ private:
 	std::vector<NovelHeader> NovelList;
 	vita2d_pgf * pgf = vita2d_load_default_pgf();	//Font!
 	int ItemSelected = 0;
-	float ScrollY = 0;
+
+	int ScrollY = 0;
 	int Scroll = 0;
 	int GridPerLine = 5;
 public:
